main buisness process completed customer,item,item waste,order,transaction

+++ src/main/java/lk/ijse/florist_pos/final_project/dto/Tm/CartTm.java
@@ -11,7 +11,7 @@ import lombok.*;
 public class CartTm {
     private String itemId;
     private String itemName;
-    private double cartQty;
+    private int cartQty;
     private double unitPrice;
     private double total;
     private String customerName;
